feat: added adaptive for section
