+++ src/components/main/main-page/MainPage.tsx
@@ -8,6 +8,7 @@ import Mission from './mission/Mission';
 import Goals from './goals/Goals';
 import Counter from './counter/Counter';
 import PressAboutUs from './press-about-us/PressAboutUs';
+import Payment from '../payment/Payment';
 import Gallery from './gallery/Gallery';
 import Partners from './partners/Partners';
 
@@ -23,6 +24,7 @@ const MainPage = () => {
         <Goals />
         <Counter />
         <PressAboutUs />
+        <Payment />
         <Gallery />
         <Partners />
       </div>
