+++ src/components/main/payment/payment-form/payment-currency/PaymentCurrency.tsx
@@ -1,0 +1,46 @@
+import { useState } from 'react';
+
+const CURRENCY = [
+  {
+    text: '₴',
+    value: 'uah',
+  },
+  {
+    text: '$',
+    value: 'usd',
+  },
+  {
+    text: '€',
+    value: 'eur',
+  },
+];
+
+const PaymentCurrency = () => {
+  const [selectedCurrency, setSelectedCurrency] = useState<string>('uah');
+
+  return (
+    <div className="flex gap-2">
+      {CURRENCY.map((item) => (
+        <div className="relative" key={item.value}>
+          <input
+            checked={selectedCurrency === item.value}
+            value={item.value}
+            type="radio"
+            name="currency-type"
+            id={item.value}
+            className="peer absolute left-0 top-0 h-0 w-0 opacity-0"
+            onChange={(e) => setSelectedCurrency(e.target.value)}
+          />
+          <label
+            htmlFor={item.value}
+            className="flex h-12 w-12 cursor-pointer items-center justify-center rounded-full bg-lightVioletSecond text-xl font-bold leading-[1] text-textViolet transition-all duration-300 ease-linear peer-checked/:bg-mainViolet peer-checked/:text-white"
+          >
+            {item.text}
+          </label>
+        </div>
+      ))}
+    </div>
+  );
+};
+
+export default PaymentCurrency;
